backend/requirements.txt
Flask
Flask-Cors
PyJWT
sentence-transformers
torch
numpy
scikit-learn
python-docx
pdfminer.six
nltk
